Move the TextureVertexData struct to the new Velaptor.OpenGL.GPUData interface & folder
