+++ Testing/VelaptorTests/OpenGL/OpenGLExtensionMethodsTests.cs
@@ -7,6 +7,7 @@ namespace VelaptorTests.OpenGL
     using System.Drawing;
     using System.Numerics;
     using Velaptor.OpenGL;
+    using Velaptor.OpenGL.GPUData;
     using Xunit;
 
     /// <summary>
