+++ Velaptor/OpenGL/Buffers/TextureGPUBuffer.cs
@@ -14,6 +14,7 @@ namespace Velaptor.OpenGL.Buffers
     using Velaptor.Graphics;
     using Velaptor.NativeInterop.OpenGL;
     using Velaptor.OpenGL.Exceptions;
+    using Velaptor.OpenGL.GPUData;
     using Velaptor.Reactables.Core;
     using Velaptor.Reactables.ReactableData;
     using NETRect = System.Drawing.Rectangle;
